Make ceph-common aware of osd config fragments
This removes the implicit order requirement when using OSD fragments.
When you use OSD fragments and ceph-osd role is not the last one,
the fragments get removed from ceph.conf by ceph-common.

It is not nice to have this code at two locations, but this is
necessary to prevent problems, when ceph-osd is the last role as
ceph-common gets executed before ceph-osd.

This could be prevented when ceph-common would be explicitly called
at the end of the playbook.

This is a backport of commit 349b9ab3e72f188c50b5f84f3ce5ee2fddcfbf80
onto Stable-2.1. That commit was by Christian Zunker
<christian.zunker@codecentric.de>

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>